# docs/api.yaml
openapi: 3.1.0
info:
  title: Parade State Management System API
  version: 0.1.0
  description: |
    API for the Battalion Parade State Management System.

    ## Authentication (issue 31)
    All endpoints (except `/auth/*`) derive the caller's identity from the
    authenticated session — the HttpOnly `session_token` cookie set by the
    Google OAuth callback, or the same token as a Bearer header for
    non-browser clients. No endpoint accepts identity from query parameters
    or request bodies; provenance (`created_by`, `updated_by`, …) is stamped
    server-side from the session user. Unauthenticated calls return 401;
    calls below the endpoint's role tier or outside the caller's scope
    grants return 403. Role is re-read from the database on every request,
    so role changes take effect immediately.

    ## Access control
    Two enforcement dimensions applied on every data endpoint:
    - **Row visibility**: user sees only personnel rows covered by their nominal-roll subunit assignments.
    - **Column visibility**: response includes only columns whose sensitivity label is met by the user's access level.
      The `columns` manifest on list endpoints is always pre-scoped — clients render headers from the manifest, never from hardcoded column names.

    Admin and super_admin roles bypass both dimensions and see all rows and columns.

    ## Column manifest pattern (Option C)
    Endpoints returning personnel data include a `columns` array alongside `rows`.
...

...
servers:
...
paths:
...
  /discussions/posts:
    get:
      tags: [discussions]
      summary: List board posts
      description: |
        Any admin/super-admin. Flat newest-first list capped at the 200
        most recent posts; optional category and status filters.
      parameters:
        - name: category
          in: query
          schema: { type: string, enum: [requests, bugs] }
        - name: status
          in: query
          schema: { type: string, enum: [Open, Duplicate, Accepted, Implemented, Closed] }
      responses:
        '200':
          description: List of posts (with author_name and comment_count)
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      tags: [discussions]
      summary: Create a board post
      description: |
        Any admin/super-admin. Category is required (requests/bugs); the
        post starts with status Open — status moves only through triage.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, body, category]
              properties:
                title: { type: string, maxLength: 200 }
                body: { type: string }
                category: { type: string, enum: [requests, bugs] }
      responses:
        '201': { description: Created }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '422': { $ref: '#/components/responses/BadRequest' }
  /discussions/posts/{postId}:
    parameters:
      - name: postId
        in: path
        required: true
        schema: { type: string }
    get:
      tags: [discussions]
      summary: Fetch a post with its comments
      responses:
        '200': { description: Post detail (comments oldest-first) }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [discussions]
      summary: Edit a post (title/body)
      description: |
        Author only (403 otherwise, super-admins included). Sets
        edited_at. Category/status are not editable here — see triage.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title: { type: string, maxLength: 200, nullable: true }
                body: { type: string, nullable: true }
      responses:
        '200': { description: Updated }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [discussions]
      summary: Delete a post and its comments
      description: Super-admin only (403 otherwise).
      responses:
        '200': { description: Deleted }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /discussions/posts/{postId}/triage:
    parameters:
      - name: postId
        in: path
        required: true
        schema: { type: string }
    patch:
      tags: [discussions]
      summary: Triage a post (category/status)
      description: |
        Super-admin only (403 otherwise). The one board action that
        writes an audit row (entity_type discussion_post); a call that
        changes nothing writes none.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                category: { type: string, enum: [requests, bugs], nullable: true }
                status:
                  type: string
                  nullable: true
                  enum: [Open, Duplicate, Accepted, Implemented, Closed]
      responses:
        '200': { description: Updated }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /discussions/posts/{postId}/comments:
    parameters:
      - name: postId
        in: path
        required: true
        schema: { type: string }
    post:
      tags: [discussions]
      summary: Comment on a post
      description: Any admin/super-admin.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [body]
              properties:
                body: { type: string }
      responses:
        '201': { description: Created }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /discussions/comments/{commentId}:
    parameters:
      - name: commentId
        in: path
        required: true
        schema: { type: string }
    patch:
      tags: [discussions]
      summary: Edit a comment (body)
      description: Author only (403 otherwise). Sets edited_at.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [body]
              properties:
                body: { type: string }
      responses:
        '200': { description: Updated }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [discussions]
      summary: Delete a comment
      description: Super-admin only (403 otherwise).
      responses:
        '200': { description: Deleted }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }

  # ===========================================================================
  # Taggings (super-admin only) — overlay of person→subunit remappings on an NR
  # ===========================================================================
...
components:
...
  responses:
    Unauthorized:
      description: No valid session
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Insufficient access level or role
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found (or not visible to this user)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    BadRequest:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'

  # ---------------------------------------------------------------------------
  # Schemas
  # ---------------------------------------------------------------------------
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true

    # --- Nominal Roll ---
...